Fixed aria-describedby issue on funding app project title view
This commit fixes a syntax issue on the funding application
project title page, where aria-describedby was incorrectly
referred to as  aria-described-by.

Additionally, this commit captures a couple of other changes;
replacing span tags with div tags for the page hint (this
was previously a span in the GOV.UK Design System guidance,
but has since changed to a div), and fixing a minor issue
where the input field class contained a space at the end.