Fixed potential XSS security vulnerability in pre-application journey
This commit fixes a potential XSS security vulnerability in the
pre-application journey. Where we were referencing a model attribute
in an interpolated string including a line break and calling the
.html_safe method, this opened up a potential avenue of attack.

A malicious actor could have inserted HTML into an organisation's
address, which would then have been rendered on the pages modified
in this commit.